Extract CardImage
