+++ front/app/components/IdeaCard/CardImage/ImagePlaceholder.tsx
